+++ backend/database/core.py
@@ -1,5 +1,8 @@
 """This file defines the database connection, plus some terminal commands for
 setting up and tearing down the database.
+
+Do not import anything directly from `backend.database._core`. Instead, import
+from `backend.database`.
 """
 import click
