+++ backend/tests/README.md
@@ -1,0 +1,5 @@
+To run tests locally:
+
+```shell
+python -m pytest
+```
