+++ requirements/_core.in
@@ -1,12 +1,14 @@
-flask
-flask-sqlalchemy
-flask-serialize
-pandas
-requests
-pytest
-pytest-cov
-PyYAML
 black
 flake8
+flask
+flask-serialize
+flask-sqlalchemy
+gunicorn
+pandas
 pip-tools
+pytest
+pytest-cov
+pytest-flask-sqlalchemy
 python-dotenv
+PyYAML
+requests
