+++ requirements/dev_windows.txt
@@ -26,11 +26,15 @@ flake8==3.8.4
 flask-serialize==1.5.0
     # via -r requirements/_core.in
 flask-sqlalchemy==2.4.4
-    # via -r requirements/_core.in
+    # via
+    #   -r requirements/_core.in
+    #   pytest-flask-sqlalchemy
 flask==1.1.2
     # via
     #   -r requirements/_core.in
     #   flask-sqlalchemy
+gunicorn==20.1.0
+    # via -r requirements/_core.in
 idna==2.10
     # via requests
 iniconfig==1.1.1
@@ -48,7 +52,9 @@ mypy-extensions==0.4.3
 numpy==1.20.1
     # via pandas
 packaging==20.9
-    # via pytest
+    # via
+    #   pytest
+    #   pytest-flask-sqlalchemy
 pandas==1.2.3
     # via -r requirements/_core.in
 pathspec==0.8.1
@@ -71,10 +77,16 @@ pyparsing==2.4.7
     # via packaging
 pytest-cov==2.11.1
     # via -r requirements/_core.in
+pytest-flask-sqlalchemy==1.0.2
+    # via -r requirements/_core.in
+pytest-mock==3.5.1
+    # via pytest-flask-sqlalchemy
 pytest==6.2.2
     # via
     #   -r requirements/_core.in
     #   pytest-cov
+    #   pytest-flask-sqlalchemy
+    #   pytest-mock
 python-dateutil==2.8.1
     # via pandas
 python-dotenv==0.15.0
@@ -90,7 +102,9 @@ requests==2.25.1
 six==1.15.0
     # via python-dateutil
 sqlalchemy==1.3.23
-    # via flask-sqlalchemy
+    # via
+    #   flask-sqlalchemy
+    #   pytest-flask-sqlalchemy
 toml==0.10.2
     # via
     #   black
@@ -106,3 +120,4 @@ werkzeug==1.0.1
 
 # The following packages are considered to be unsafe in a requirements file:
 # pip
+# setuptools
